Rework redirect, make banditTest, rework template

+++ src/Models/BanditTest.php
@@ -5,24 +5,25 @@ namespace Georges\Models;
 
 class BanditTest
 {
-
     public $test;
     public array $variations;
     public array $goals;
-
+    public array $visits;
 
     public function __construct()
     {
         $this->test        = new Test();
         $this->variations   = [];
         $this->goals        = [];
+        $this->visits        = [];
     }
 
-    public function setBanditTest(Test $test, array $variations, array $goals)
+    public function setBanditTest(Test $test, array $variations, array $goals, array $visits)
     {
         $this->test        = $test;
         $this->variations   = $variations;
         $this->goals        = $goals;
+        $this->visits        = $visits;
     }
 
     public function getBanditTestById(int $id)
@@ -30,6 +31,7 @@ class BanditTest
         $testManager = new Test();
         $variationManager = new Variation();
         $goalManager = new Goal();
+        $VisitManager = new VisitLog();
 
         $test = $testManager->get($id);
 
@@ -38,14 +40,26 @@ class BanditTest
         if (!empty($this->test)) {
             $listVariation = $variationManager->getList(500, ['name' => 'test_id', 'agrega' => '=', 'value' => $id]);
 
+            if (!empty($listVariation)) {
+                foreach ($listVariation as $oneVariation) {
+                    $variation = new Variation();
+                    array_push($this->variations, $variation->hydrate($oneVariation));
 
-            foreach ($listVariation as $oneVariation) {
-                $variation = new Variation();
-                array_push($this->variations, $variation->hydrate($oneVariation));
-                $listGoal = $goalManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $oneVariation['variation_id']]);
-                foreach ($listGoal as $oneGoal) {
-                    $goal = new Goal();
-                    array_push($this->goals, $goal->hydrate($oneGoal));
+                    $listGoal = $goalManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $oneVariation['variation_id']]);
+                    if (!empty($listGoal)) {
+                        foreach ($listGoal as $oneGoal) {
+                            $goal = new Goal();
+                            array_push($this->goals, $goal->hydrate($oneGoal));
+                        }
+                    }
+
+                    $listVisit = $VisitManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $oneVariation['variation_id']]);
+                    if (!empty($listVisit)) {
+                        foreach ($listVisit as $oneVisit) {
+                            $visit = new VisitLog();
+                            array_push($this->visits, $visit->hydrate($oneVisit));
+                        }
+                    }
                 }
             }
         }
@@ -54,13 +68,12 @@ class BanditTest
 
     public function searchBanditTest(string $uri = "/", array $filter = null, string $statut = "all")
     {
-        $listVariationsTest = [];
-        $listGoalsTest = [];
         $listBanditTest = [];
 
         $testManager = new Test();
         $variationManager = new Variation();
         $goalManager = new Goal();
+        $VisitManager = new VisitLog();
 
         $tests = $testManager->getList(
             10000,
@@ -72,28 +85,44 @@ class BanditTest
         );
 
         foreach ($tests as $oneTest) {
-            $listVariation = $variationManager->getList(500, ['name' => 'test_id', 'agrega' => '=', 'value' => $oneTest['test_id']]);
+            $listVariationsTest = [];
+            $listGoalsTest = [];
+            $listVisitsTest = [];
 
             $test = new Test();
             $oneTest = $test->hydrate($oneTest);
 
-            foreach ($listVariation as $oneVariation) {
-                $variation = new Variation();
-                array_push($listVariationsTest, $variation->hydrate($oneVariation));
+            $listVariation = $variationManager->getList(500, ['name' => 'test_id', 'agrega' => '=', 'value' => $oneTest->testId]);
 
-                //Get Goal List
-                $listGoal = $goalManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $oneVariation['variation_id']]);
-                foreach ($listGoal as $oneGoal) {
-                    $goal = new Goal();
-                    array_push($listGoalsTest, $goal->hydrate($oneGoal));
+            if (!empty($listVariation)) {
+                foreach ($listVariation as $oneVariation) {
+                    $variation = new Variation();
+                    array_push($listVariationsTest, $variation->hydrate($oneVariation));
+
+                    //Get Goal List
+                    $listGoal = $goalManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $variation->variationId]);
+                    if (!empty($listGoal)) {
+                        foreach ($listGoal as $oneGoal) {
+                            $goal = new Goal();
+                            array_push($listGoalsTest, $goal->hydrate($oneGoal));
+                        }
+                    }
+
+                    //Visi List
+                    $listVisit = $VisitManager->getList(100000, ['name' => 'variation_id', 'agrega' => '=', 'value' => $variation->variationId]);
+                    if (!empty($listVisit)) {
+                        foreach ($listVisit as $oneVisit) {
+                            $visit = new VisitLog();
+                            array_push($listVisitsTest, $visit->hydrate($oneVisit));
+                        }
+                    }
                 }
             }
 
             $bandit = new BanditTest();
-            $bandit->setBanditTest($oneTest, $listVariationsTest, $listGoalsTest);
+            $bandit->setBanditTest($oneTest, $listVariationsTest, $listGoalsTest, $listVisitsTest);
             array_push($listBanditTest, $bandit);
         }
-
         // return BanditTest array matching uri_regex,filter and statut
         return $listBanditTest;
     }
@@ -109,20 +138,32 @@ class BanditTest
             $goal->save();
         }
 
+        foreach ($this->visits as $visit) {
+            $visit->save();
+        }
+
         return true;
     }
 
     public function delete()
     {
         $test_id = $this->test->testId;
-        $this->test->delete(6);
+        $this->test->delete($test_id);
 
-        foreach ($this->variations as $variation) {
-            $variation->delete($variation->variationId);
+        if (!empty($this->variations)) {
+            foreach ($this->variations as $variation) {
+                $variation->delete($variation->variationId);
+            }
         }
-
-        foreach ($this->goals as $goal) {
-            $goal->delete($goal->goalId);
+        if (!empty($this->goals)) {
+            foreach ($this->goals as $goal) {
+                $goal->delete($goal->goalId);
+            }
+        }
+        if (!empty($this->visits)) {
+            foreach ($this->visits as $visit) {
+                $visit->delete($visit->visitLogId);
+            }
         }
 
         return true;
